+++ ledger/participant-integration-api/src/test/suite/scala/platform/apiserver/services/ApiSubmissionServiceSpec.scala
@@ -376,7 +376,7 @@ class ApiSubmissionServiceSpec
       )(any[LoggingContext])
     ).thenReturn(Future.unit)
 
-    CurrentLedgerConfiguration
+    IndexStreamingCurrentLedgerConfiguration
       .owner(
         configManagementService,
         config = LedgerConfiguration(
